+++ packages/ui-components/src/dropdown-menu/dropdown-menu.tsx
@@ -126,7 +126,6 @@ export type MenuItem = {
     preventClose?: boolean;
     before?: React.ReactNode;
     after?: React.ReactNode;
-    title?: string;
 };
 
 export type DropdownMenuProps = {
@@ -180,7 +179,6 @@ export const Dropdown = (props: DropdownProps): JSX.Element => {
                                     }}
                                     role="menuitem"
                                     style={item.style}
-                                    title={item.title}
                                 >
                                     {item.before}
                                     <ItemLabel>